+++ pkg/saas/clusters.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package v1alpha1
+package saas
 
 import (
 	batchv1 "k8s.io/api/batch/v1"
