+++ pkg/saas/clusters_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package v1alpha1
+package saas
 
 import (
 	"encoding/json"
